requirements-test.txt
molecule>=24.2.0
molecule-plugins[docker]>=23.5.3
docker>=7.0.0
pytest-testinfra>=10.0.0
